move TODOs out of docs, remain limiter_model
